backend/pyproject.toml
[project]
name = "supoclip-backend"
version = "0.1.0"
description = "Python-based backend for SupoClip"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "openai-whisper>=20250625",
    "pydantic-ai>=0.4.9",
    "setuptools-rust>=1.11.1",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2025.7.21",
    "whisper>=1.1.10",
    "greenlet>=2.0.0",
    "moviepy>=2.2.1",
    "opencv-python>=4.8.0",
]

[tool.uv]
package = true
